requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
langchain-huggingface
ipykernel
pyttsx3
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
tf-keras
faiss-cpu
chromadb
langchain-openai
streamlit
langchain_groq
fastapi
uvicorn
langserve
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
langchain
gradio
crewai
crewai_tools
transformers
langchain-huggingface
sentence-transformers
neo4j
